Implement new BarChart
