+++ src/components/BarChart/__tests__/index.tsx
@@ -1,4 +1,4 @@
-import { render, waitFor, screen } from "@testing-library/react";
+import { render, screen } from "@testing-library/react";
 import React from "react";
 import { build, fake, bool } from "@jackfranklin/test-data-bot";
 import BarChart, { Props, DataPoint } from "..";
